feat(stripe): Reader strategy
