+++ providers/stripe/internal/reader/custom.go
@@ -1,4 +1,4 @@
-package stripe
+package reader
 
 import (
 	"github.com/amp-labs/connectors/internal/jsonquery"
