+++ providers/stripe/internal/reader/test/balance_transactions/nested-source-nested-payment-intent.json
