+++ providers/stripe/internal/reader/test/balance_transactions/nested-source.json
