+++ providers/stripe/internal/reader/test/checkout-sessions/with-line-items.json
