+++ providers/stripe/internal/reader/test/customers/1-first-page.json
