+++ providers/stripe/internal/reader/test/customers/with-metadata.json
