+++ providers/stripe/internal/reader/test/general-errors/error-bad-request.json
