+++ providers/stripe/metadata.go
@@ -10,5 +10,5 @@ import (
 func (c *Connector) ListObjectMetadata(
 	ctx context.Context, objectNames []string,
 ) (*common.ListObjectMetadataResult, error) {
-	return metadata.Schemas.Select(c.Module(), objectNames)
+	return metadata.Schemas.Select(common.ModuleRoot, objectNames)
 }
