+++ providers/stripe/write.go
@@ -13,7 +13,7 @@ func (c *Connector) Write(ctx context.Context, config common.WriteParams) (*comm
 		return nil, err
 	}
 
-	url, err := c.getURL(config.ObjectName)
+	url, err := c.Base.GetURL(config.ObjectName)
 	if err != nil {
 		return nil, err
 	}
